gstreamer_interface refactored.

+++ generate_uml.sh
@@ -1,5 +1,6 @@
 clang-uml
+[ -d "puml" ] && echo "Directory puml exists." || mkdir puml
 cd puml/
-rm MultiRelay_class_diagram.png
+[ -e MultiRelay_class_diagram.png ] && rm MultiRelay_class_diagram.png
 plantuml MultiRelay_class_diagram.puml
 cd ..
